POTENTIAL BREAKING CHANGE: feat: allow fully Lua `packer_compiled` file (#407)
* feat(packer): provide default for `compile_to_lua`

This new option determines whether packer should wrap lua in a vim file.
In the case users do not specify, it will be `false`.

* ref(packer): pass `output_to_lua` to `compile`

* ref(packer): autodetect `compile_lua` setting

It should be possible to determine whether the user wants to compile a
Lua file based on the output path. If the user wants to output a lua
file, then we compile to lua. If the user wants to output to vim, then
compile to vim.

* ref(compile): `compile_to_lua` -> `compile_lua`

* fix(packer): use `:luaf` when sourcing lua files

Previously, when only Vim files were supported, only the `:source`
command was needed. Now, since some of the files will be Lua, there has
to be a check to determine whether we should use `:source` for Vim or
`:luaf` for Lua.

* feat(compile): add lua templates

* feat(compile): implement lua compilation

* docs: setting `compile_path` to a Lua file

* style(compile): use `~=` instead of `not ==`

* perf(packer): use `:source` for both Lua and VimL

Newer commits allow `:source` to be used for both Lua and VimL files.
This was an unecessary distinction on my part.

* ref(packer): change `output_path` to Lua by default

This change makes `:PackerCompile` generate a Lua file by default.

BREAKING CHANGE: users manually sourcing their 'packer_compiled.vim' may
                 find it is not working as intended, unless they have
                 specified a custom `output_path` which points to a Vim
                 file.

* docs: update default `packer_compiled` path

After the previous commit, the default `packer_compiled` was changed
from a VimL file to a Lua file. Reflecting this change, the documentation
has been updated to show the new path.

Additionally, since discussion on the PR for this feature has indicated
that Vim output will go away eventually, I have removed advertising of
that feature (i.e. I did not put "-- can also be a '.vim' file").

RELATED: https://github.com/wbthomason/packer.nvim/pull/407#pullrequestreview-693533235
RELATED: https://github.com/wbthomason/packer.nvim/pull/407#discussion_r659460782

* feat(packer): remove VimL packer_compiled if `output_lua`

This commit adds functionality which removes a user's old
'packer_compiled' VimL file if they compile it with Lua. It does so by
looking for a file of the same name but different extension (e.g.
'packer_compiled.lua' & 'packer_compiled.vim') and removes the VimL
variant. This only occurs if the files are in the same directory, so if
a user alters their `compile_path` when changing to Lua it will not be
detected (e.g. 'plugin/packer_compiled.vim' -> 'plugin/packer.lua' won't
be caught).

* ref(packer): warn when removing `packer_compiled`

It was discussed on the PR that we should warn users after removing
their old `packer_compiled` files, since they may not be on a version of
Neovim which can automatically source them.

* Fix conditional and run lua-format

* fix: handle case of switching filetype back to vim

This still doesn't handle people changing their compile path to
something other than the same prefix with a different extension, but it
will account for both directions of extension change.

Co-authored-by: Wil Thomason <wil.thomason@gmail.com>